[AI] refactor: extract model-runtime scope resolver to its own file (#120154)
